Generated Classes for Dep Module Monomorphizations Use Shorter Hashed Name
Previously the entire unique module name which encodes the full path to the module's Bazel target definition was used for the name of dep module monomorphizations. This ensured a certainty that the generated classes were uniquely associated with the module in question. Unfortunately, this led to the situation where javac was trying to generate files of the same name, which ran into "File name too long" errors from the OS itself. This would've led to limitations on the nesting level of the file structure of a Claro project, which is obviously untenable. Now, the unique module name is hashed using sha256 to produce a class/file name whose max length is determined by the length of the module target name and the procedure name. There's now a minuscule possibility of collisions in generated monomorphization class names.