Implement Auto-Generated Docs to Validate Example Claro Programs in Docs
This is going to be a huge help to ensure that the docs site (https://jasonsteving99.github.io/claro-lang/) actually represents the latest state of Claro, and the example programs are always actually runnable and emitting the claimed output.

Updates the GitHub Workflow that builds new releases of the docs site to actually run the docs generation to validate docs are all as they should be before deploying.