Convert @jzwood's codegen.sh Tool Into create_claro_project.sh
This new approach creates a single script that will be able to set a new user up with all the files they need to have a fully functioning Claro project up and running with all the Bazel config files necessary and not requiring manually copying files from examples/bzlmod/ (which a few users have already been confused by).

This new script is auto-generated from the files in examples/bzlmod/ so that I don't have to keep both in sync. This does mean, however, that I'll need to auto-generate a release artifact with this built CLI tool so that users can click a button to download the script and get running.