Add Support for Missing Primitive Types: `long`, `double`, `char`
These are all just oversights that should've always existed but I moved past during the very early stages of Claro development. These should all be properly supported now including basic mathematical operator support and auto-promotion (not on function calls which could arguably be a nice feature add in keeping with the Java universe way of things).

Still need to come back and change the return type of exponentiation and log to return `double` instead of `float`.

I also need to change the Java representation for the Claro `float` type from `java.lang.Double` to `java.lang.Float`. Left it as is for now so that I can start by getting the bootstrapping compiler to acknowledge the new types.