Codegen Concrete Classes for Every Concrete Struct Type
Claro's builtin struct type follows this idea of being "polymorphic" in the sense that `struct {foo: int, bar: string}` can just be referenced inline as if it just exists already. Before this change, under the hood all structs were represented at runtime with the single $ClaroStruct Java type where there was more or less an array of Object such that field accesses were codegen'd into something like ((ExpectedType) s.values[i]) which is quite ridiculous. It additionally created additional runtime overhead (casting) on every single struct field access when this is obviously not necessary. This new approach avoids all of that overhead and should have the added benefit of making it much less likely to run into any compiler bugs where Java codegen is missing necessary casts.

It'll take an update to the bootstrapping compiler for dep module monomorphization to operate properly with this new approach so this is a breaking change.