Complete Migration of All Structs to Use Generated Monomorphic Java Class
This finally drops the previously used polymorphic ClaroStruct.java to represent all structs language-wide using essentially `Object[]` and casts for field reads.

The power of this new approach will be fully realized once ALL builtin structured types are migrated to this same approach. Without getting into too many details, the short story is that once all builtin structured types are represented monomorphically at compile-time, then the JVM bytecode generated should be statically guaranteed to be able to bypass dynamic dispatch calls to do essentially dynamic vtable lookups to find which method to dispatch to, and instead, all calls should be able to use the fast path to directly call the single static function that could possibly be referenced. This will also ensure that the JVM's JIT is able to very easily perform optimizations in places where they wouldn't be possible in vanilla Java where every method on a value with a non-final class would have to be dynamically dispatched as it's not known whether the value is actually of the named type or of a subtype.