Drop concrete class codegen (#41)
Unfortunately, there is not currently time to fully commit going down this road of migrating Claro's nested types to be codegen'd to use concrete classes that fully specify their runtime types. As I'm reaching the end of my year of full time work on Claro this is just a more pervasive change than I actually have time to work through for now. At the moment it's a much higher priority to truly work through documenting the current state of Claro than trying to work through this more idealized approach to more efficient runtime characteristics.

# Oversight in Original Planning
I can't just leave things partially migrated to using this approach, as this is something of an all or nothing change. In particular, Claro's type system is designed to allow the following:
```
var nestedOneof: tuple<oneof<int, string>, int> = (10, 10); 
```
Currently, Claro's unable to correctly codegen the rhs initialization as tuple<oneof<int, string>, int> and will instead codegen as tuple<int, int> because the LHS asserted type is not being used to override the implied type of the RHS for the sake of codegen. This is certainly something that can be worked around, but in my estimation would (in combination with the remaining migration of struct, list, map and user-defined-type) take so much time that I would lose my opportunity to put sufficient effort into documentation that this project needs more than anything at this point in time.

# Maintaining history
I'm intentionally reverting these commits in order to maintain the history so that I can potentially revisit these changes to make redoing this at some point in the future easier.

-------------------------------------------------

* Revert "Complete Migration of All Structs to Use Generated Monomorphic Java Class"

This reverts commit e74fd2aa237024847f9bb26f29d9f3332edc8989.

* Revert "Codegen Concrete Classes for Every Concrete Struct Type"

This reverts commit eb488e14e30651d2e689335577f4ceb7b2b960a5.

* Revert "Codegen Concrete Classes for Every Concrete Struct Type"

This reverts commit cffc5f56dff03e909179e999df02e6059655a957.

* Revert "Codegen Concrete Classes for Every Concrete Struct Type"

This reverts commit 39faa6c788bb5cb97a640c192873998c43512c22.

* Revert "Codegen Concrete Classes for Every Concrete Struct Type"

This reverts commit 4f7ec92de099b54ba01767c75fc7e15bb558e7c5.

* Revert "Codegen Concrete Classes for Every Concrete Struct Type"

This reverts commit d65970ffd28e03a3fcfc540429fa01e903395d45.

* Revert "Codegen Concrete Classes for Every Concrete Struct Type"

This reverts commit 54f09ddd82bc5645c2006978954e48628276ba81.