Update Auto Validated Docs Formatting of Examples
This just makes the templating syntax standardized so that you just need to drop in smth like `{{EX1}}` on its own line and you'll get the code block inside a divided section autogenerated for you. Also now separates the code and output into their own code blocks to make it easier to read than the prior commented out section in the same codeblock as the src.