Add CNAME Record for docs.clarolang.com
This allows docs.clarolang.com to be directed to the GH Pages site. Originally this was only included on the gh-pages branch, but then the GH Actions workflow that releases new versions of the site was going to overwrite that branch with the contents of the main branch and effectively delete the CNAME file. So now, it's included in the main branch to avoid deleting it.