Prototype a Better ClaroDocs Site Using React/Node/Vite
This is currently just a prototype that's hardcoding some fake module dependency data. But conceptually this should already be able to run over arbitrary Claro programs as long as its dep graph can be converted into the appropriate data structure (currently found in `//tools/clarodocs/src/claro_module_apis/demo.tsx`).

The hope is that this can eventually be used to produce an autogenerated site for arbitrary `claro_binary()` targets that updates automatically as modules are edited. It's still TBD whether that will actually be feasible (especially when referenced by downstream Bazel projects) - but it seems like it should be possible.