Update the Prototype of a Better ClaroDocs Site Using React/Node/Vite
Update to actually use real dep graph and module apis of a real Claro program! This uncovered various "parsing" bugs in HighlightJS.tsx so now this solution should be relatively solid for arbitrary Claro programs going forward.

The hope is that this can eventually be used to produce an autogenerated site for arbitrary `claro_binary()` targets that updates automatically as modules are edited. It's still TBD whether that will actually be feasible (especially when referenced by downstream Bazel projects) - but it seems like it should be possible.