Add a Better Auto Docs Site Generator for Arbitrary claro_binary() Targets (#53)
This v2 of the auto docs generator is much more visually appealing and has more functionality than the previous. Unfortunatley, I still haven't figured out how to enable auto-reloading of the site when the dep graph changes in some way, so I really want to continue investigating that feature as that's what would really make this useful. But for now, this is at least a great way to do some static inspection of large Claro programs.