Support Browser History in ClaroDocs Generated Site
This was a fairly straightforward change that stores state in URL query params instead of using React's `useState()` directly since now when the site auto-reloads, it'll reload to the specific URL that it was at before the reload and so you won't get disrupted in your browsing of the site just because you made a code change. There're probably some weird edge cases here where you'd need to manually navigate back to the DepGraph (home) if the query params are invalid, but that's fine. Arguably, this should now be in a good enough state to actually push a release to the BCR with this.