Prepare to Push to BCR
Last bit of housecleaning before attempting to push this version to the BCR. Primarily just updates examples/bzlmod/ to use clarodocs().

Updates the JS related Module deps to not be considered "dev" dependencies as they need to be available for downstream Modules depending on claro-lang.

Also fixes the stdlib/cache module so that it correctly builds.